protocol: add color-management-v2.xml
This does not implement the protocol, it will simply run the scanner
on the xml to generate the .c and .h files.

This is a variant of the color-management protocol that still didn't
land upstream on the wayland-protocols repository.

We are using that to experiment with the protocol, but we'll
completely drop this eventually, when the official protocol
lands.

Signed-off-by: Leandro Ribeiro <leandro.ribeiro@collabora.com>